[WIP] refactor
