+++ app/src/main/java/com/softyorch/taskapp/data/network/pexels/PexelsService.kt
@@ -1,7 +1,7 @@
 package com.softyorch.taskapp.data.network.pexels
 
 import com.softyorch.taskapp.utils.DataOrError
-import com.softyorch.taskapp.data.network.pexels.responseMyCollection.Media
+import com.softyorch.taskapp.data.network.pexels.response.Media
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import retrofit2.Retrofit
