+++ app/src/main/java/com/softyorch/taskapp/data/network/pexels/response/PexelsMyCollections.kt
@@ -1,4 +1,4 @@
-package com.softyorch.taskapp.data.network.pexels.responseMyCollection
+package com.softyorch.taskapp.data.network.pexels.response
 
 data class PexelsMyCollections(
     val id: String,
