+++ app/src/main/java/com/softyorch/taskapp/data/network/pexels/response/VideoPicture.kt
@@ -1,4 +1,4 @@
-package com.softyorch.taskapp.data.network.pexels.responseMyCollection
+package com.softyorch.taskapp.data.network.pexels.response
 
 data class VideoPicture(
     val id: Int,
