+++ app/src/main/java/com/softyorch/taskapp/domain/pexelUseCase/GetImage.kt
@@ -1,7 +1,7 @@
 package com.softyorch.taskapp.domain.pexelUseCase
 
 import com.softyorch.taskapp.utils.DataOrError
-import com.softyorch.taskapp.data.network.pexels.responseMyCollection.Media
+import com.softyorch.taskapp.data.network.pexels.response.Media
 import com.softyorch.taskapp.data.repository.PexelsRepository
 
 class GetImage(private val repository: PexelsRepository) {
